refactor: useCvc를 엔티티 레이어로 이동 및 핸들러 훅 구현
