+++ src/core/hooks/useInput.ts
@@ -1,9 +1,5 @@
 import { useState } from 'react';
 
-interface UseInputProps {
-  validator: (value: string) => boolean;
-}
-
 export interface UseInputResult {
   value: string;
   touched: boolean;
@@ -11,12 +7,11 @@ export interface UseInputResult {
   handleBlur: () => void;
 }
 
-export const useInput = ({ validator }: UseInputProps): UseInputResult => {
+export const useInput = () => {
   const [value, setValue] = useState<string>('');
   const [touched, setTouched] = useState<boolean>(false);
 
   const handleChange = (inputValue: string): void => {
-    if (inputValue !== '' && !validator(inputValue)) return;
     setValue(inputValue);
     setTouched(false);
   };
