+++ src/entities/card/cvc/model.ts
@@ -5,7 +5,10 @@ const ERROR_MESSAGE = {
   RANGE: 'CVC를 전부 채워주세요.',
 };
 
-export const validateCvc = (cvc: string): string | undefined => {
+export const getErrorCvc = (cvc: string): string | undefined => {
   if (cvc.length !== CVC_LENGTH) return ERROR_MESSAGE.RANGE;
   return;
+};
+export const validateCvc = (cvc: string): boolean => {
+  return getErrorCvc(cvc) !== undefined;
 };
