+++ src/entities/card/index.ts
@@ -1,0 +1,2 @@
+export * from './cvc/useCvc';
+export * from './cvc/model';
